feature: display tree in grid
Add commands `get_relationships` and `get_person_layers`.
Generate grid layout on basis of the person layers.
Add `GridItem` that contains either a `Person` to be used in a
`PersonCard` or a `Connection[]` to be used in a `Connector`.
The `Connector` component will be used to display the various
relationship and children lines between the nodes.