feat: add wasm support
In addition to tauri, the app can now run solely in the browser with wasm.
Add library baumstamm-wasm that implements the commands and is compilable to wasm.
Rename baumstamm-app to baumstamm-tauri.
Extract grid logic into separate baumstamm-grid library, which can be used by tauri and wasm.
Call wasm commands from the app api, when in wasm context.
Adapt svelte app to the differences between the targets.
Add vite plugins to support wasm.